# backend/swagger.yaml
swagger: "2.0"
info:
  description: "APIs of Logistic Web Service"
  version: "1.0.0"
  title: "Logistic Web Service"
  contact:
    email: "d.bed11@gptgdt.net"
host: "localhost:8000"
basePath: "/api"
schemes:
- "http"
- "https"
...
paths:
...
  /feature:
    get:
      tags:
        - "feature"
      summary: get feature information
      produces:
        - application/json
      parameters:
        - in: query
          name: pageSize
          type: string
          description: limit number of features on a page
          parameters:
        - in: query
          name: page
          type: string
          description: num of page
        - in: query
          name: keyword
          type: string
          description: search by keyword
        - in: query
          name: name
          type: string
          description: filter by name
        - in: query
          name: detail
          type: string
          description: filter by detail 
        - in: query
          name: sortBy
          type: string
          description: sort in order (name, -updatedAt)
      responses:
        200:
          description: get feature information successfully.
          schema:
            type: object
            example: {
                  "name": "string",
                  "logo": "path",
                  "detail": "string",
            }
        400:
          description: career information is not found.  
  /feature/{featureId}:
    get:
      tags:
        - "feature"
      summary: get feature by id
      produces:
        - application/json
      parameters:
        - in: path
          name: featureId
          type: string
          required: true
          description: Id of a feature 
      responses:
        200:
          description: get feature successfully.
          schema:
            type: object
            example: {
              "name": "string",
              "logo": "path",
              "detail": "string",              
            }
        400:
          description: feature is not found.
  /feature/service/{serviceId}:
    get:
      tags:
        - "feature"
      summary: get features by service id
      produces:
        - application/json
      parameters:
        - in: query
          name: pageSize
          type: string
          description: limit number of features on a page
          parameters:
        - in: query
          name: page
          type: string
          description: num of page
        - in: path
          name: serviceId
          type: string
          required: true
          description: Id of a service 
      responses:
        200:
          description: get feature successfully.
          schema:
            type: object
            example: {
              "name": "string",
              "logo": "path",
              "detail": "string",              
            }
        400:
          description: feature is not found.
  /admin/feature/{serviceId}:
    post:
      tags:
        - "feature"
      summary: "Create a new feature"
      consumes:
        - multipart/form-data
      parameters:
        - in: path
          name: serviceId
          schema: 
            type: string
          required: true
          description: service id to add feature in
        - in: body
          name: feature
          description: a new feature.
          schema:
            type: object
            required:
              - name
              - detail
            properties:
              name:
                type: string
              detail:
                type: string  
        - in: formData
          name: logo
          type: file
          description: a logo for feature.
          required: true
      requestBody:
        content: 
          multipart/form-data:
            schema:            
              type: object
              properties:      
                name:           
                  type: string
                detail:   
                  type: string
                logo: 
                  type: string
                  format: binary
      responses:
        200:
          description: Create feature successfully.
          schema:
            type: object
            example: {
              "name": "string",
              "logo": "path",
              "detail": "string",
            }
        400:
          description: Create feature failed.
      security:
        - bearerAuth: []
  /admin/feature/{featureId}:
    put:
      tags:
        - "feature"
      summary: "Update a existing feature"
      consumes:
        - multipart/form-data
      parameters:
        - in: path
          name: featureId
          schema: 
            type: string
          required: true
          description: feature id
        - in: body
          name: feature
          description: a existing feature.
          schema:
            type: object
            required:
              - name
              - detail
            properties:
              name:
                type: string
              detail:
                type: string
        - in: formData
          name: logo
          type: file
          description: a logo for feature.
          required: true
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - detail
              properties:
                name:
                  type: string
                logo:
                  type: string
                  format: base64 
                detail:
                  type: string
      responses:
        200:
          description: Update feature successfully.
          schema:
            type: object
            example: {
              "name": "string",
              "logo": "path",
              "detail": "string",
            }
        400:
          description: Update feature failed.
      security:
        - bearerAuth: []
    delete:
      tags:
        - "feature"
      summary: "Delete a existing feature"
      produces:
        - application/json
      parameters:
        - in: path
          name: featureId
          type: string
          required: true
          description: Id of a feature 
      responses:
        200:
          description: Delete feature successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        400:
          description: Update feature failed.
...
  /distance:
    get:
      tags:
        - "distance"
      summary: get distance information
      produces:
        - application/json
      parameters:
        - in: query
          name: pageSize
          type: string
          description: limit number of distances on a page
          parameters:
        - in: query
          name: page
          type: string
          description: num of page
        - in: query
          name: fromProvince
          type: string
          description: Origin province
          required: true
        - in: query
          name: toProvince
          type: string
          description: Destination province
          required: true
      responses:
        200:
          description: get distance information successfully.
          schema:
            type: object
            example: {
                  "fromProvince": "string",
                  "toProvince": "path",
                  "zonecode": "string",
                  "distance":  "number",
            }
        400:
          description: distance information is not found.  
  /distance/{distanceId}:
    get:
      tags:
        - "distance"
      summary: get distance by id
      produces:
        - application/json
      parameters:
        - in: path
          name: distanceId
          type: string
          required: true
          description: Id of a distance 
      responses:
        200:
          description: get distance successfully.
          schema:
            type: object
            example: {
              "fromProvince": "string",
              "toProvince": "path",
              "zonecode": "string",
              "distance":  "number",            
            }
        400:
          description: distance is not found.
    put:
      tags:
        - "distance"
      summary: "Update an existing distance zonecode"
      consumes:
        - application/json
      parameters:
        - in: path
          name: distanceId
          schema: 
            type: string
          required: true
          description: distance id
        - in: body
          name: distance
          description: an existing distance.
          schema:
            type: object
            required:
              - zonecode
            properties:              
              zonecode:
                type: string
      responses:
        200:
          description: Update distance successfully.
          schema:
            type: object
            example: {
              "fromProvince": "string",
              "toProvince": "path",
              "zonecode": "string",
              "distance":  "number",
            }
        400:
          description: Update distance failed.
      security:
        - bearerAuth: []
  /distance/service/{serviceId}:
    get:
      tags:
        - "distance"
      summary: get distances by service id
      produces:
        - application/json
      parameters:
        - in: query
          name: pageSize
          type: string
          description: limit number of distances on a page
          parameters:
        - in: query
          name: page
          type: string
          description: num of page
        - in: path
          name: serviceId
          type: string
          required: true
          description: Id of a service 
      responses:
        200:
          description: get distance successfully.
          schema:
            type: object
            example: {
              "fromProvince": "string",
              "toProvince": "path",
              "zonecode": "string",
              "distance":  "number",          
            }
        400:
          description: distance is not found.
...
  /admin/distance/{distanceId}:
    put:
      tags:
        - "distance"
      summary: "Update an existing distance"
      consumes:
        - application/json
      parameters:
        - in: path
          name: distanceId
          schema: 
            type: string
          required: true
          description: distance id
        - in: body
          name: distance
          description: an existing distance.
          schema:
            type: object
            required:
              - fromProvince
              - toProvince
              - zonecode
              - dist
            properties:
              fromProvince:
                type: string
              toProvince:
                type: string
              zonecode:
                type: string
              dist:
                type: number
      responses:
        200:
          description: Update distance successfully.
          schema:
            type: object
            example: {
              "fromProvince": "string",
              "toProvince": "path",
              "zonecode": "string",
              "distance":  "number",
            }
        400:
          description: Update distance failed.
      security:
        - bearerAuth: []
    delete:
      tags:
        - "distance"
      summary: "Delete an existing distance"
      produces:
        - application/json
      parameters:
        - in: path
          name: distanceId
          type: string
          required: true
          description: Id of a distance 
      responses:
        200:
          description: Delete distance successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    description: Success message
        400:
          description: Delete distance failed.
      security:
        - bearerAuth: []
...
  /service/{serviceId}:
    get:
      tags:
        - "service"
      summary: get delivery service by id
      produces:
        - application/json
      parameters:
        - in: path
          name: serviceId
          type: string
          required: true
          description: Id of a service 
      responses:
        200:
          description: get service successfully.
          schema:
            type: object
            example: {
              "name": "string",
              "sub_detail": "string",    
              "target": "string",   
              "tip" : "string",
              "quotes": ["quote1", "quote2"],
              "logo": "path",
              "banner": "path",
              "features": ["feature1", "feature2"],    
              "participants": ["participant1", "participant2"],     
              "price": "string",  
              "distances": ["distance1", "distance"],   
              "price_files": ["path1", "path2"],          
            }          
        400:
          description: service is not found.    
...
securityDefinitions:
  bearerAuth:
    type: "apiKey"
    name: "Authorization"
    in: "header"